split schema files
