+++ app/drizzle/schema/schema.tsx
@@ -1,0 +1,3 @@
+import { pgSchema } from "drizzle-orm/pg-core";
+
+export const moonsyncSchema = pgSchema("moonsync");
